add zeropad2d v0.3

+++ python/paddle/fluid/tests/unittests/test_zeropad2d.py
@@ -104,7 +104,7 @@ class TestZeroPadOp2d(unittest.TestCase):
         expect_res = np.pad(
             x, [[0, 0], [0, 0], [pad[2], pad[3]], [pad[0], pad[1]]])
 
-        x_tensor = to_tensor(x)
+        x_tensor = to_tensor(x, dtype='int32')
         pad_tensor = to_tensor(pad)
         ret_res = zeropad2d(x_tensor, pad_tensor).numpy()
         self.assertTrue(np.allclose(expect_res, ret_res))
